# backend/internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NexusHub API
  version: 0.1.0
  description: |
    REST API for NexusHub — a WireGuard VPN management dashboard with
    PostgreSQL, JWT auth, and eBPF-backed security rules.

    **Conventions**

    - All endpoints live under `/api/v1`.
    - Authenticated routes require `Authorization: Bearer <access_token>`.
    - List endpoints accept `?limit=`, `?offset=`, and `?sort=` (prefix
      with `-` for descending) and return a `PageEnvelope`.
    - Errors use the `Error` schema with a stable `code` field — rely on
      that rather than on `error` copy.
    - Every response echoes `X-Request-ID`. Clients may send their own
      (must be a UUID) or let the server generate one. `X-Correlation-ID`
      is only echoed when the caller supplied it.
  contact:
    name: NexusHub
    url: https://github.com/tomeksdev/NexusHub
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Relative to the deployment base URL
...
security:
  - bearerAuth: []
paths:
...
  /peers/{id}:
    parameters:
      - $ref: '#/components/parameters/IDPath'
    get:
      tags: [Peers]
      summary: Get a peer by ID
      responses:
        '200':
          description: Peer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Peer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Peers]
      summary: Delete a peer
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IDPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message. Do not pattern-match on this.
        code:
          type: string
          description: Stable machine-readable code. Safe to branch on.
          enum:
            - INVALID_REQUEST
            - INVALID_CREDENTIALS
            - ACCOUNT_LOCKED
            - ACCOUNT_DISABLED
            - UNAUTHORIZED
            - FORBIDDEN
            - REFRESH_INVALID
            - REFRESH_REUSED
            - RATE_LIMITED
            - NOT_FOUND
            - CONFLICT
            - IP_POOL_EXHAUSTED
            - INTERNAL
...
    Peer:
      type: object
      required:
        - id
        - interface_id
        - name
        - public_key
        - allowed_ips
        - assigned_ip
        - dns
        - status
        - rx_bytes
        - tx_bytes
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        interface_id: { type: string, format: uuid }
        owner_user_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        public_key: { type: string }
        allowed_ips:
          type: array
          items: { type: string }
        assigned_ip: { type: string }
        endpoint: { type: string, nullable: true }
        persistent_keepalive: { type: integer, nullable: true }
        dns:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, disabled, expired]
        last_handshake: { type: string, format: date-time, nullable: true }
        rx_bytes: { type: integer, format: int64 }
        tx_bytes: { type: integer, format: int64 }
        expires_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...